extract useCreatePolicyAction hook to deduplicate create-policy logic
- Extract shared create-policy logic (modal state, RBAC check, create
  mutation, policy-name dedup) into useCreatePolicyAction hook
- Refactor PoliciesList and SinglePolicyView to use the shared hook
- Use proper Route[] type for SinglePolicyView props
- Add comment clarifying error fallback in MultiplePoliciesView
